Added some changes in networklight primitive and added some more adjustments in _deleteMachineNetwork function

+++ packages/grid_client/src/high_level/base.ts
@@ -7,6 +7,7 @@ import { events } from "../helpers/events";
 import { Operations, TwinDeployment } from "../high_level/models";
 import { DeploymentFactory } from "../primitives/deployment";
 import { Network } from "../primitives/network";
+import { ZNetworkLight } from "../primitives/networklight";
 import { Nodes } from "../primitives/nodes";
 import { Deployment } from "../zos/deployment";
 import { Workload, WorkloadTypes } from "../zos/workload";
@@ -27,6 +28,7 @@ class HighLevelBase {
       WorkloadTypes.ip,
       WorkloadTypes.ipv4, // TODO: remove deprecated
       WorkloadTypes.zmachine,
+      WorkloadTypes.zmachinelight,
       WorkloadTypes.zmount,
       WorkloadTypes.volume,
       WorkloadTypes.zdb,
@@ -38,7 +40,9 @@ class HighLevelBase {
   ): [Workload[], Workload[]] {
     let deletedMachineWorkloads: Workload[] = [];
     if (names.length === 0) {
-      deletedMachineWorkloads = deployment.workloads.filter(item => item.type === WorkloadTypes.zmachine);
+      deletedMachineWorkloads = deployment.workloads.filter(
+        item => item.type === WorkloadTypes.zmachine || item.type === WorkloadTypes.zmachinelight,
+      );
     }
 
     if (names.length !== 0 && types.includes(WorkloadTypes.zmachine)) {
@@ -64,9 +68,33 @@ class HighLevelBase {
       }
     }
 
+    if (names.length !== 0 && types.includes(WorkloadTypes.zmachinelight)) {
+      console.log("In here");
+      const Workloads = deployment.workloads.filter(item => item.type === WorkloadTypes.zmachinelight);
+      for (const workload of Workloads) {
+        if (!names.includes(workload.name)) {
+          continue;
+        }
+        for (const mount of workload.data["mounts"]) {
+          names.push(mount.name);
+        }
+
+        const toRemoveZlogs = deployment.workloads.filter(item => {
+          const x = item.type === WorkloadTypes.zlogs;
+          const y = (item.data as any)["zmachine-light"] === workload.name;
+          return x && y;
+        });
+
+        names.push(...toRemoveZlogs.map(x => x.name));
+
+        names.push(workload.data["network-light"].public_ip);
+        deletedMachineWorkloads.push(workload);
+      }
+    }
+
     const remainingWorkloads: Workload[] = [];
     for (const workload of deployment.workloads) {
-      if (workload.type === WorkloadTypes.network) {
+      if (workload.type === WorkloadTypes.network || workload.type === WorkloadTypes.networklight) {
         remainingWorkloads.push(workload);
         continue;
       }
@@ -86,67 +114,104 @@ class HighLevelBase {
     remainingWorkloads: Workload[],
     deletedMachineWorkloads: Workload[],
     node_id: number,
-  ): Promise<[TwinDeployment[], Workload[], number[], string[], Network | null]> {
+  ): Promise<[TwinDeployment[], Workload[], number[], string[], Network | ZNetworkLight | null]> {
     const twinDeployments: TwinDeployment[] = [];
     const deletedNodes: number[] = [];
     const deletedIps: string[] = [];
     const deploymentFactory = new DeploymentFactory(this.config);
-    let network: Network | null = null;
+    let network: Network | ZNetworkLight | null = null;
+    let deletedIp;
+    let contract_id;
+
+    let numberOfIps;
     for (const workload of deletedMachineWorkloads) {
       if (!network) {
         const networkName = workload.data["network"].interfaces[0].network;
         const networkIpRange = Addr(workload.data["network"].interfaces[0].ip).mask(16).toString();
-        network = new Network(networkName, networkIpRange, this.config);
+        if (workload.type === WorkloadTypes.zmachinelight) {
+          network = new ZNetworkLight(networkName, networkIpRange, this.config);
+        } else {
+          network = new Network(networkName, networkIpRange, this.config);
+        }
         await network.load();
       }
       const machineIp = workload.data["network"].interfaces[0].ip;
       events.emit("logs", `Deleting ip: ${machineIp} from node: ${node_id}, network ${network.name}`);
-      const deletedIp = network.deleteReservedIp(node_id, machineIp);
+      deletedIp = network.deleteReservedIp(node_id, machineIp);
       if (remainingWorkloads.length === 0) {
         twinDeployments.push(new TwinDeployment(deployment, Operations.delete, 0, 0, "", network));
       }
-      const numberOfIps = network.getNodeReservedIps(node_id).length;
+      numberOfIps = network.getNodeReservedIps(node_id).length;
       if (numberOfIps !== 0) {
         console.log(`network ${network.name} still has ${numberOfIps} ip(s) reserved`);
         deletedIps.push(deletedIp);
         continue;
       }
-      const hasAccessPoint = network.hasAccessPoint(node_id);
-      if (hasAccessPoint && network.nodes.length !== 1) {
-        console.log(
-          `network ${network.name} still has access point:${hasAccessPoint} and number of nodes ${network.nodes.length}`,
-        );
-        deletedIps.push(deletedIp);
-        continue;
+      if (network instanceof Network) {
+        const hasAccessPoint = network.hasAccessPoint(node_id);
+        if (hasAccessPoint && network.nodes.length !== 1) {
+          console.log(
+            `network ${network.name} still has access point:${hasAccessPoint} and number of nodes ${network.nodes.length}`,
+          );
+          deletedIps.push(deletedIp);
+          continue;
+        }
       }
+      if (network instanceof Network) {
+        contract_id = await network.deleteNode(node_id);
+      } else if (network instanceof ZNetworkLight) {
+        const contracts = await network.getDeploymentContracts(network.name);
+        contract_id = contracts[0].contractID;
+      }
+      console.log("contract_id", contract_id);
+      console.log("remainingWorkloads in delete", remainingWorkloads);
+      console.log("deployment.contract_id", deployment.contract_id);
 
-      const contract_id = await network.deleteNode(node_id);
       if (contract_id === deployment.contract_id) {
+        console.log("in hereee contract_id", contract_id);
         if (remainingWorkloads.length === 1) {
           twinDeployments.push(new TwinDeployment(deployment, Operations.delete, 0, 0, "", network));
           remainingWorkloads = [];
         } else {
+          console.log("Remaining workloads in else", remainingWorkloads);
           remainingWorkloads = remainingWorkloads.filter(item => item.name !== network?.name);
           deletedIps.push(deletedIp);
           deletedNodes.push(node_id);
         }
       } else {
+        console.log("outer else");
+        if (network instanceof ZNetworkLight) {
+          twinDeployments.push(new TwinDeployment(deployment, Operations.delete, 0, 0, "", network));
+        }
+        console.log("network.deployments", network.deployments);
+        console.log("network.", network);
+
         // check that the deployment doesn't have another workloads
         for (let d of network.deployments) {
+          console.log("network.deployments", d);
+
           d = await deploymentFactory.fromObj(d);
           if (d.contract_id !== contract_id) {
             continue;
           }
           if (d.workloads.length === 1) {
+            console.log("outer else workloads =1");
+
             twinDeployments.push(new TwinDeployment(d, Operations.delete, 0, 0, "", network));
           } else {
+            console.log("outer else not workloads =1");
+
             d.workloads = d.workloads.filter(item => item.name !== network?.name);
             twinDeployments.push(new TwinDeployment(d, Operations.update, 0, 0, "", network));
           }
         }
       }
       // in case of the network got more accesspoints on different nodes this won't be valid
-      if (network.nodes.length === 1 && network.getNodeReservedIps(network.nodes[0].node_id).length === 0) {
+      if (
+        network instanceof Network &&
+        network.nodes.length === 1 &&
+        network.getNodeReservedIps(network.nodes[0].node_id).length === 0
+      ) {
         const contract_id = await network.deleteNode(network.nodes[0].node_id);
         for (let d of network.deployments) {
           d = await deploymentFactory.fromObj(d);
@@ -172,6 +237,7 @@ class HighLevelBase {
       WorkloadTypes.ip,
       WorkloadTypes.ipv4, // TODO: remove deprecated
       WorkloadTypes.zmachine,
+      WorkloadTypes.zmachinelight,
       WorkloadTypes.zmount,
       WorkloadTypes.volume,
       WorkloadTypes.zdb,
@@ -181,7 +247,7 @@ class HighLevelBase {
       WorkloadTypes.zlogs,
     ],
   ): Promise<TwinDeployment[]> {
-    if (types.includes(WorkloadTypes.network)) {
+    if (types.includes(WorkloadTypes.network) || types.includes(WorkloadTypes.networklight)) {
       throw new GridClientErrors.Workloads.WorkloadDeleteError("Network workload can't be deleted.");
     }
     let twinDeployments: TwinDeployment[] = [];
@@ -201,7 +267,11 @@ class HighLevelBase {
       await this._deleteMachineNetwork(deployment, remainingWorkloads, deletedMachineWorkloads, node_id);
     twinDeployments = twinDeployments.concat(newTwinDeployments);
     remainingWorkloads = newRemainingWorkloads;
-    if (remainingWorkloads.length !== 0 && remainingWorkloads.length < numberOfWorkloads) {
+    if (
+      network instanceof Network &&
+      remainingWorkloads.length !== 0 &&
+      remainingWorkloads.length < numberOfWorkloads
+    ) {
       for (const deleteNode of deletedNodes) {
         await network!.deleteNode(deleteNode);
       }
